// frontend/pizzaiolo/src/app/model/pizzaiolo/orderEditDTO.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CestaEditDTO } from './cestaEditDTO';

export interface OrderEditDTO { 
    address?: string;
    amount?: number;
    comment?: string;
    idOrder?: number;
    idUser?: string;
    orderDate?: Date;
    /**
     * Lista de identificadores de pizzas.
     */
    pizzas?: Array<CestaEditDTO>;
}