// frontend/pizzaiolo/src/app/model/pizzaiolo/orderStatusEditDTO.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface OrderStatusEditDTO { 
    deliveryDate?: Date;
    idChef?: string;
    idCourier?: string;
    idOrder?: number;
    idUser?: string;
    /**
     * Estados de pedido.
     */
    orderStatus?: OrderStatusEditDTO.OrderStatusEnum;
}
export namespace OrderStatusEditDTO {
    export type OrderStatusEnum = 'cancelado' | 'elaborandose' | 'enviado' | 'preparado' | 'recibido' | 'solicitado';
    export const OrderStatusEnum = {
        Cancelado: 'cancelado' as OrderStatusEnum,
        Elaborandose: 'elaborandose' as OrderStatusEnum,
        Enviado: 'enviado' as OrderStatusEnum,
        Preparado: 'preparado' as OrderStatusEnum,
        Recibido: 'recibido' as OrderStatusEnum,
        Solicitado: 'solicitado' as OrderStatusEnum
    };
}